+++ CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [0.9.5] 2020
+### Features
+- new custom action (**+**) to open the system backups (snapshots). It is just a shortcut to Timeshift.
+    <p align="center">
+        <img src="https://raw.githubusercontent.com/vinifmor/bauh/staging/pictures/releases/0.9.5/backup_action.png">
+    </p>
 
 ### Improvements
 - Arch